Moving installers to Internal namespace so that they do not clutter the super namespace which should be used for actual tests
