+++ Neo4j.Driver/Neo4j.Driver.IntegrationTests/Internals/INeo4jInstaller.cs
@@ -1,8 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using System.Threading.Tasks;
 
-namespace Neo4j.Driver.IntegrationTests
+namespace Neo4j.Driver.IntegrationTests.Internals
 {
   /// <summary>
   /// Neo4j installer
